Remove global scope checks from permissions-api
Roles are to be tenant-scoped by design for MVP. We can consider
adding global scope in the future, but for now to keep things simple
we're just going to have tenant roles.
