+++ internal/api/permissions.go
@@ -8,39 +8,6 @@ import (
 	"go.infratographer.com/permissions-api/internal/query"
 )
 
-func (r *Router) checkGlobalScope(c *gin.Context) {
-	scope := c.Param("scope")
-
-	ctx, span := tracer.Start(c.Request.Context(), "api.checkGlobalScope")
-	defer span.End()
-
-	actor, err := currentActor(c)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error(), "message": "failed to get the actor"})
-		return
-	}
-
-	actorResource, err := query.NewResourceFromURN(actor.urn)
-	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"message": "error processing actor URN", "error": err.Error()})
-		return
-	}
-
-	err = query.ActorHasGlobalPermission(ctx, r.authzedClient, actorResource, scope, "")
-	if err != nil {
-		if errors.Is(err, query.ErrScopeNotAssigned) {
-			c.JSON(http.StatusForbidden, gin.H{"message": "actor does not have requested scope"})
-			return
-		}
-
-		c.JSON(http.StatusInternalServerError, gin.H{"message": "an error occurred checking actor scopes", "error": err.Error()})
-
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{})
-}
-
 func (r *Router) resourcesAvailable(c *gin.Context) {
 	ctx, span := tracer.Start(c.Request.Context(), "api.resourcesAvailable")
 	defer span.End()
