+++ internal/query/stores.go
@@ -1,0 +1,9 @@
+package query
+
+import "github.com/authzed/authzed-go/v1"
+
+// Stores represents a SpiceDB store.
+type Stores struct {
+	SpiceDB       *authzed.Client
+	SpiceDBPrefix string
+}
